Organize contrib & add disclaimer
